requirements.txt
altair==4.2.2
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.1
streamlit==1.21.0
streamlit_option_menu==0.3.2
openpyxl==3.0.9